requirements.txt
tensorflow
numpy
matplotlib
pandas
jinja2
gin-config